requirements.txt
appdirs==1.4.4
APScheduler==3.9.1.post1
argcomplete==2.0.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-browser-reload==1.6.0
django-tailwind==3.4.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyserial==3.5
python-escpos==2.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typed-ast==1.5.4
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
whitenoise==6.2.0
wrapt==1.14.1